fix circular import
Changes:
- fix circular import between `matable_fit` and `matable_fit_cas2022`.
- Update tests from `romanisim` to work completely within `stcal`.
- Allow test to use `romanisim` if present.
