+++ src/stcal/ramp_fitting/matable_fit.py
@@ -34,60 +34,7 @@ import numpy as np
 from scipy import interpolate
 
 from . import matable_fit_cas2022
-
-
-# Read Time in seconds
-#   For Roman, the read time of the detectors is a fixed value and is currently
-#   backed into code. Will need to refactor to consider the more general case.
-#   Used to deconstruct the MultiAccum tables into integration times.
-READ_TIME = 3.04
-
-
-def ma_table_to_tbar(ma_table, read_time=READ_TIME):
-    """Construct the mean times for each resultant from an ma_table.
-
-    Parameters
-    ----------
-    ma_table : list[list]
-        List of lists specifying the first read and the number of reads in each
-        resultant.
-
-    Returns
-    -------
-    tbar : np.ndarray[n_resultant] (float)
-        The mean time of the reads of each resultant.
-    """
-    firstreads = np.array([x[0] for x in ma_table])
-    nreads = np.array([x[1] for x in ma_table])
-    meantimes = read_time * firstreads + read_time * (nreads - 1) / 2
-    # at some point I need to think hard about whether the first read has
-    # slightly less exposure time than all other reads due to the read/reset
-    # time being slightly less than the read time.
-    return meantimes
-
-
-def ma_table_to_tau(ma_table, read_time=READ_TIME):
-    """Construct the tau for each resultant from an ma_table.
-
-    .. math:: \\tau = \\overline{t} - (n - 1)(n + 1)\\delta t / 6n
-
-    following Casertano (2022).
-
-    Parameters
-    ----------
-    ma_table : list[list]
-        List of lists specifying the first read and the number of reads in each
-        resultant.
-
-    Returns
-    -------
-    :math:`\\tau`
-        A time scale appropriate for computing variances.
-    """
-
-    meantimes = ma_table_to_tbar(ma_table)
-    nreads = np.array([x[1] for x in ma_table])
-    return meantimes - (nreads - 1) * (nreads + 1) * read_time / 6 / nreads
+from .matable_util import ma_table_to_tau, ma_table_to_tbar
 
 
 def construct_covar(read_noise, flux, ma_table):
@@ -241,167 +188,6 @@ def ki_and_variance_grid(ma_table, flux_on_readvar_pts):
         outki.append(ki)
         outvar.append(variances)
     return np.array(outki), np.array(outvar)
-
-
-# okay, how do we get what we want?  we want to input a bunch of
-# read_noise and fluxes, and get out a bunch of ki and outvar.
-# it feels like we can just set up a bunch of 1D interpolators?
-
-class RampFitInterpolator:
-    """Ramp fitting tool aiding efficient fitting of large number of ramps.
-
-    The basic idea is that for a given image, ignoring cosmic rays or saturated
-    pixels, the ramp fitting parameters are just a linear combination of the
-    resultants.  The weights of this linear combination are a single parameter
-    family in the flux in the ramp divided by the read variance.  So rather than
-    explicitly calculating those weights for each pixel, we can up front calculate
-    them overa  grid in the flux over the read variance, and interpolate off that
-    grid for each point.  That can all be done in a vectorized way, allowing one
-    to avoid doing something like a matrix inverse for each of a 16 million pixels.
-
-    The tool pre-calculates the grid and interpolators it needs at initialization,
-    and then uses the results of that calculation when invoked to get the weights
-    :math:`k` or variances.  The expectation is that most users just initialize
-    and then call the fit_ramps method.
-    """
-    def __init__(self, ma_table, flux_on_readvar_pts=None):
-        """Construct a RampFitInterpolator for an ma_table and a grid of
-        flux/read_noise**2.
-
-        Parameters
-        ----------
-        ma_table : list[list] (int)
-            list of lists of first reads, number of reads in each resultant
-        flux_on_readvar_pts : np.ndarray (float)
-            flux / read_noise**2 points at which to compute ramp fitting
-            matrices.
-            if None, a default grid will be used that should cover all reasonable
-            values, from the read variance being 100k larger to 100k smaller than
-            the electrons per second.
-        """
-        if flux_on_readvar_pts is None:
-            self.flux_on_readvar_pts = 10.**(
-                np.linspace(-5, 5, 200, dtype='f4'))
-        else:
-            self.flux_on_readvar_pts = flux_on_readvar_pts
-        ki, var = ki_and_variance_grid(ma_table, self.flux_on_readvar_pts)
-        self.ma_table = ma_table
-        self.ki_vals = ki
-        self.var_vals = var
-        self.ki_interpolator = interpolate.CubicSpline(
-            self.flux_on_readvar_pts, self.ki_vals)
-        self.var_interpolator = interpolate.CubicSpline(
-            self.flux_on_readvar_pts, self.var_vals)
-
-        # interp1d has a lot of options, but the timing difference on my system
-        # is swamped by the cpu boosting up or down?
-        # most of the time in this function goes to evaluating the interpolators,
-        # so it's worth playing around a little to see if there's a win there.
-        # self.ki_interpolator = interpolate.interp1d(
-        #     self.flux_on_readvar_pts, self.ki_vals, axis=0)
-        # self.var_interpolator = interpolate.interp1d(
-        #     self.flux_on_readvar_pts, self.var_vals, axis=0)
-
-    def ki(self, flux, read_noise):
-        """Compute :math:`k`, the weights for the linear combination of
-        resultant differences for optimal measurement of ramp pedestal and
-        slope.
-
-        Parameters
-        ----------
-        flux : array_like (float)
-            Estimate of electrons per second in ramp
-        read_noise : array_like (float)
-            read_noise in ramp.  Must be broadcastable with flux.
-
-        Returns
-        -------
-        ki : array_like[..., 2, n_resultant] (float)
-            :math:`k`, weights of differences in linear combination of
-            ramp pixels
-        """
-        # clip points outside range to edges.
-        fluxonreadvar = flux / read_noise**2
-        if isinstance(fluxonreadvar, u.Quantity):
-            unit = fluxonreadvar.unit
-        else:
-            unit = 1
-        fluxonreadvar = np.clip(
-            fluxonreadvar, self.flux_on_readvar_pts[0] * unit,
-            self.flux_on_readvar_pts[-1] * unit)
-
-        return self.ki_interpolator(fluxonreadvar).astype('f4')
-
-    def variances(self, flux, read_noise):
-        """Compute the variances of ramp fit parameters.
-
-        Parameters
-        ----------
-        flux : array_like (float)
-            Estimate of electrons per second in ramp
-        read_noise : array_like (float)
-            read_noise in ramp.  Must be broadcastable with flux.
-
-        Returns
-        -------
-        variances : array_like[..., 3, 2, 2] (float)
-            covariance of ramp fit parameters, for read noise, poisson noise,
-            and the total noise
-        """
-        # clip points outside range to edges.
-        fluxonreadvar = flux / read_noise**2
-        if isinstance(fluxonreadvar, u.Quantity):
-            unit = fluxonreadvar.unit
-        else:
-            unit = 1
-        fluxonreadvar = np.clip(
-            fluxonreadvar, self.flux_on_readvar_pts[0] * unit,
-            self.flux_on_readvar_pts[-1] * unit)
-        var = self.var_interpolator(fluxonreadvar).astype('f4')
-        read_noise = np.array(read_noise)
-        read_noise = read_noise.reshape(
-            read_noise.shape + (1,) * (len(var.shape) - len(read_noise.shape)))
-        var *= read_noise**2
-        return var
-
-    def fit_ramps(self, resultants, read_noise, fluxest=None):
-        """Fit ramps for a set of resultants and their read noise.
-
-        Does not handle partial ramps (i.e., broken due to CRs).
-
-        Parameters
-        ----------
-        resultants : np.ndarray[n_resultants, nx, ny] (numeric)
-            Resultants to fit
-        read_noise : float or array_like like resultants
-            read noise in array
-        fluxest : float or array_like like resultants
-            Initial estimate of flux in each ramp, in electrons per second.
-            If None, estimated from the median flux differences between
-            resultants.
-
-        Returns
-        -------
-        par : np.ndarray[nx, ny, 2] (float)
-            the best fit pedestal and slope for each pixel
-
-        var : np.ndarray[nx, ny, 3, 2, 2] (float)
-            the covariance matrix of par, for each of three noise terms:
-            the read noise, Poisson source noise, and total noise.
-        """
-        tbar = ma_table_to_tbar(self.ma_table)
-        dtbar = np.diff(tbar).astype('f4')
-        differences = resultants_to_differences(resultants)
-        if fluxest is None:
-            dtbar_reshape = dtbar.reshape(
-                (dtbar.shape[0],) + (1,) * len(differences.shape[1:]))
-            fluxest = np.median(differences[1:] / dtbar_reshape, axis=0)
-        if (np.ndim(fluxest) == 0) and (np.ndim(resultants) > 1):
-            fluxest = fluxest * np.ones(resultants.shape[1:])
-        ki = self.ki(fluxest, read_noise)
-        par = np.einsum('...cd,d...->...c', ki, differences)
-        var = self.variances(fluxest, read_noise)
-        return par, var
 
 
 def resultants_to_differences(resultants):
